pyani_graphics.py: Accounted for matplotlib display problem
issue#7 on GitHub was a problem with matplotlib, deriving from the
difference between max and min values in the heatmap colourbar being
zero. This change forces a minimum difference of 0.01, avoiding the
matplotlibe issue. This is not a breaking bug in matplotlib 1.3.1, but
is known to occur in 1.4.2 and 1.4.3, breaking the script.