genbank_get_genomes_by_taxon.py: now handles nuccore_assembly_refseq
Previous versions of the script relied on the availability of
nuccore_assembly_insdc, which it turns out can't be guaranteed.
In its absence, nuccore_assembly_refseq works in about the same
way. However, the final variant I've found -
nuccore_assembly_wgsmaster (in the absence of either of the other
versions) is still not working.