genbank_get_genomes_by_taxon.py: updated for difficult NCBI downloads
A number of circumstances could occur that made the previous version of
this script fail to download a genome. These revisions fix all the
problems I was aware of.

i) Elink could return links from an assembly UID to contig UIDs in
three ways: assembly_nuccore_insdc (most cases); assembly_nuccore_refseq
(most of the others); and assembly_nuccore_wgsmaster. If the refseq
or insdc links were available, the download could proceed as normal.
With wgsmaster-only links, there is no way to use Elink to get the
contig data, so the script guesses at/munges together a URL that should
point to an archive of the contigs in a FASTA file, and downloads these
directly.

ii) If Elink needs to return links to over 100000 contigs (which can
happen), then these are capped at 100000 and I could not see how to
extend the returned list beyond this. In this circumstance, the
wgsmaster archive is downloaded instead.

iii) Genomes with many contigs (>10000) were prone to repeated
download failure, so these are now batched.