BLAST methods now write output to subfolder
A similar change for the BLAST-based methods ANIb and ANIblastall
is implemented: comparison output is now written to a subfolder
of the output directory. This output is compressed to a .tar.gz
file and the subdirectory deleted, unless the --nocompress
option is passed (this option is new to this commit).