Modified NCBI genome download script to use new NCBI FTP locations
I was receiving informal notices of download errors with the
`genbank_get_genomes_by_taxon.py` script. Some of this was genuine
bugs in the script - issues with the various ways I was munging
downloads depending how dat was available, and some was Python
API inconsistencies (download metadata in httplib!).

To bypass most of this, the script now downloads assemblies from
the common ftp://ftp.ncbi.nlm.nih.gov/genomes/all/ location at NCBI,
using information from the Entrez eSummary. This is NOT foolproof.
In testing on Phytophthora, at least two genomes are not accessible
in this way - but at least the script is now graceful enough to
skip those that can't be downloaded, and report which assembly
UIDs were skipped, when it's done.