Fixed ANIm hanging with divergent input sequences
After fixing the ANIb empty BLAST issue (#27), further testing
revealed a bug where ANIm analysis would hang at the nucmer runs.
Manual running of the command-lines generated empty output, but
went to completion. The bug was reminiscent of a similar occurrence
in the differential primer finding code, that I had fixed with an
alternative method of calling applications with multiprocessing,
to avoid use of a callback function.

The changes in this commit drop the callback function, and use
multiprocessing's builtin methods to recover return values from
called tools.