restore seaborn colorbars, and seaborn plots for large datasets
The seaborn rendering was failing on two counts. Firstly, the
colorbar interface for sns.colormap had changed, and the
row_colors/col_colors pd.Series needed to carry the same
index column as the data frame used for the heatmap. Secondly, for
large datasets, the total figsize was getting so large that seaborn
was dumping the core. The heatmap_seaborn code now sets a maximum
figure size, and then tweaks the font scaling to try to maintain
an attractive image.