enable graceful replacement of classes for input not in classes.txt
These changes implement a more graceful handling of errors/
omissions in the classes.txt file. Individual input sequences that
are not in the classes.txt file are now treated as single classes
for the colour bars. Also, the average_nucleotide_identity.py
script now warns if there are misformatted lines in classes.txt