fix issue #91 - incorrect coverage and %ID with ANIm
This issue arose because with NUCmer alignment some regions were
being aligned twice or more (e.g. repeats) and all alignments
contributed to the ANIm calculations. This is fixed here by
adding a delta-filter -1 step to the process, to restrict each
region of the genome to participate in only a single alignment.