deprecate testing of legacy BLAST
This is the first step in completely deprecating the legacy BLAST
ANI methods. The corresponding executables are no longer available
from NCBI and, therefore, cannot be used in our continuous integration
tests.

The legacy BLAST tests are currently skipped, but will be removed,
along with the code that they operate on.