convert ANIm subcommand to use the ORM interface to pyanidb
At this point, the database is populated (with the new schema) much
as before. We still need to convert the ANIb and TETRA subcommands,
as well as all the reporting subcommands, to use the new schema/ORM.
We should also populate new tables to store results matrices for
each run.

We also need to refactor the main subcmd_anim() function