fix issue #132
TETRA analysis would fail if one or more of the input genomes did
not contain at least one representative of each of the 256 4-mer
base combinations. This would error because the check on sorted
4-mer dictionary key lengths would fail: they would be different
lengths in at least one comparison.

The fix pre-populates each dictionary with the full set of 256
4-mer base combinations, so that the test for dictionary sanity
is no longer necessary.