move old scripts under pyani/scripts and add entry points
The scripts average_nucleotide_identity.py,
delta_filter_wrapper.py and genbank_get_genomes_by_taxon.py are
moved to the same script folder as the pyani script. Rather than
being installed as script files, the entry points to these modules
are referenced.

Where useful, pyani's logging methods are substituted in for the
old arrangement (defining the logger in the script). Some elements
of the code are tidied up for pylint.