fix CircleCI/local test inconsistency
The inconsistency showed as differing alignment lengths for the same pairwise comparison,
and was fixed by enforcing that input files for ANIm analysis should be sorted when
preparing command-lines.

It seems that the outputs of A vs B and B vs A from NUCmer are not as symmetrical as
the algorithm would suggest, and it may be that we need to address this issue more
deeply.