fix issue #178
This was a knotty problem. The user had input files whose names were
all interpretable as floating-point numbers. When loading the output
CSV files, the names of files were interpreted as float64 datatypes,
but only in the dataframe index, not the headers. This meant that
the label/class files were not being used, and the sequence names
in the graphical output had extra digits where the floats did not
have an exact representation.

The solution was to read the CSV file without the `index_col=0`
argument, but specifying that column zero shoul dbe the str
datatype. Once loaded, this column could be specified as a new
index for the dataframe.

This fix forms the basis for release 0.2.10