fix graphics failure due to label/class file read errors
The bug here manifested as a failure to write graphical output
due to having the wrong number of labels (zero, as it turned out).
The source of the bug was that average_nucleotide_identity.py was
using the old label/class reader, but the new format from pyani
has three columns rather than two.