Update anim.py
The original script does not parse delta files correctly. In seven column lines, column 5 and 6 are equal when using nucmer, so the following lines do nothing.
```
            indels = int(line[4]) - int(line[5])
            aln_length -= indels
```
Delta files code gap in the reference sequence by the number of negative numbers following the 7 column line. (Query gaps are coded by the positive numbers. The absolute value of the numbers indicate how far the gap is from either the start of the alignment or the previous gap position; distance of 1 is the first following position.)

The suggested edit corrects the alignment length summed up. The sim_errors were already correctly collected.