fix error in alignment length calculation for ANIm
To calculate alignment length, we take the length of the aligned region of
the reference (no gaps), and process the delta information. This takes the
form of one value per line, following the header sequence. Positive values
indicate an insertion in the reference; negative values a deletion in the
reference (i.e. an insertion in the query). The total length of the alignment
is then:

reference_length + insertions - deletions

For example:

A = ABCDACBDCAC$
B = BCCDACDCAC$
Delta = (1, -3, 4, 0)
A = ABC.DACBDCAC$
B = .BCCDAC.DCAC$

A is the reference and has length 11. There are two insertions (positive delta),
and one deletion (negative delta). Alignment length is then 11 + 1 = 12.

This change has a minor effect on percentage identity/coverage calculation -
results still pass concordance check against JSpecies output. Some regression
test changes were required to accommodate the new exact values.