modify expected values in test_anim_delta()
The original test still expected two values to be returned from
parse_delta() - the tuple returned now has four entries.

In addition, the content of the tuple has changed. It now returns
the count of aligned bases from reference and query sequences,
then the calculated percentage identity, and the count of
similarity errors.