Rounding Error Investigation
We are currently working on changing ANIm calculations to more closely resemble dnadiff.pl output.
Regarding the Average %ID, we have agreed to calculate the weighted %ID.

However, we have found that in certain cases our values do not match on the second decimal place.
Here, we investigated whether this could be a result of a rounding error, as dnadiff uses values
for all individual sequence identities rounded to the second decimal place.

Here, we found that although in some cases the discrepancies in the Average %ID can be due to
rounding error, in some cases they are not. We found that the %ID for individual aligned regions
differs between our reported values and the ones provided in the coords files.