Update --run_results columns
Previous plans were to add separate columns for identity and alignment
lengths for both query and reference. This is because we now:
i) calulate %ID and genome covereger diffrently, and ii) run two
comparisons, resulting in diffrent values being reported for query and
subject. We wanted to report all values, as this could be of intrest
to some. However, this will require to change internals and we agreed
to postpone it to pyani-plus.

Here, I made changed to report a single value for identity and alignment
lengths, which are reported for query rather than the subject.